Some refactor for ComposableDeviceFullInfo
